Fix map_to_s3 configuration
Environment variables are not supported by Lambda@Edge, so replace their
usage with a configuration file. A JSON format file was chosen to
prevent additional dependencies.

Also, the transformed URI is now prepended with a forward-slash, "/", to
correct the path to the file object.

Tests were updated to reflect these changes.

Fixes #8, fixes #10