Provide a top-level alias for lambda function
Previously, the relevant lambda function we want to use was located
at:

  cdn_lambda.functions.map_to_s3.map_to_s3.lambda_handler

...and this function name needs to be used when configuring the
lambda in AWS. Alternatively, one could create the deployment
package by zipping up cdn_lambda/functions/map_to_s3 rather than
the entire module and using a shorter name.

Either option there is cumbersome and leaks details about how the
modules are structured internally, which could lead to cases where
we have to make changes to the lambda function config because
we refactored some modules.

It seems more stable and maintainable if we make sure there are
top-level imports for anything we expect to be invoked as a lambda
function. We can use the convention of naming them after their
trigger to denote how they're expected to be used.

This means that the function intended to be used with origin request
events can now be accessed as:

  cdn_lambda.origin_request

...which is brief, easy to understand and easy to keep working even
if we change completely how the Python modules are arranged
internally.