docs: remove some unicode icons from diagram
Unfortunately, the chosen characters here don't render correctly
when graphviz runs on Travis CI machines.  There is probably a way
to fix that by installing more fonts, but it may not be worth
spending the time on that.

Rendering as an SVG instead would also fix this but seems to
introduce its own problems.