Implement template for Lambda@Edge deployment [RHELDST-8]
This commit introduces a template used by AWS CloudFormation to
create/update the resources for Lamda@Edge applications.

Those resources include; AWS CloudFront Distribution, AWS Lambda@Edge
functions, AWS IAM Roles for each function, and AWS CloudWatch Alarms
for each function.

The template takes a parameter for the environment -- dev, stage, or
prod -- and a parameter for the environment's origin access identity ID,
for which the exodus-storage template is responsible for creating.