Implement template for CodePipeline [RHELDST-25]
This commit introduces a template used by AWS CloudFormation to
create/update the resources for CodePipeline pipelines.

Those resources include; an Amazon SNS topic, CodePipeline pipeline, S3
bucket, CodeBuild project, and the required AWS IAM roles.

The template takes a parameter for the environment -- dev, stage, or
prod -- a parameter for the environment's origin access identity ID, a
parameter for the AWS region, and a parameter for the email address for
notifications.