Enable configuration of exodus-lambda loggers [RHELDST-662]
The loggers are configured using a config file in the deployment package.
The levels and formatting of the loggers may be configured in accordance
with the logging.config.dictConfig documentation.

By default, the current behavior of the loggers is maintained; the
loggers are configured to log messages at the warning level and higher.

Additionally, the LambdaBase class' conf_file parameter now also accepts a
dict. The user can pass in a path to the lambda's config file, or a dict
containing the lambda's config.