Correct pipeline template for stage,prod environments
Previously, the template for creating pipelines only worked for the
initial, dev pipeline. This was due to stage and prod pipelines
attempting to reference resourses only created in the dev environment.
This commit replaces those references with the reousrce ARNs.

Additionally, the final pipeline action was changed to upload the
build as a zip file, preserving the template file name. Otherwise,
CodePipeline assigns a random name to the template which is impossible
for the deploy stage to predict.