Implement GPG verification [RHELDST-687]
This commit introduces a new CodeBuild project to be deployed manually
via CloudFormation prior to the creation of CodePipelines. The project
is run by the pipeline following GitHub source stages, receiving the
git commit ID from the pipeline, and verifies that the pushed commit
has a tag that is signed by an authorized entity.

This verify-source CodeBuild project requires the following environment
variables;
- REPO_OWNER: Name of the repository's owner
- REPO_NAME: Name of the repository
- REPO_BRANCH: Name of the repository's branch
- COMMIT_ID: ID of the commit to verify