Add integration tests to pipeline [RHELDST-719]
This commit introduces an additional "Test" stage to the
stage and prod lambda pipelines.

The "Test" stage contains an "IntegrationTests" action, which
runs the integration tests against the environment's CloudFront
distribution via a CodeBuild project.

If the integration tests pass, the BuildArtifact of the successful
run is preserved in the pipeline-artifacts bucket as build-${env}.

If the integration tests fail, an SNS notification is sent and a
manual rollback is triggered. The manual rollback is triggered by
a CloudWatch Event Rule that uses a CodeBuild project as a target.
The CodeBuild project deploys the CloudFormation template from the
BuildArtifact of the last successful run.