Parameterize "exodus" in CloudFormation templates [RHELDST-1035]
This commit removes many of the "exodus" strings throughout the template
and introduces a "project" parameter to the CloudFormation templates.
Parameterizing the project's name enables creation of the multiple stacks
containing similar resources for through testing.

This commmit also includes the creation of a CloudTrail logs bucket and a
modification to the existing CloudTrail.