Resolve resource dependencies in CloudFormation templates [RHELDST-1035]
If any resources defined in a CloudFormation template have a dependency
on another resource that is defined in the same template, we can expect
the creation of any dependent resources to fail.

To ensure that all resources are created in a single deployment, we
must resolve any resource dependencies in the templates through use
of the 'DependsOn' attribute.

The addition of the 'DependsOn' attribute required some IAM roles
to be created in additional stacks. This ensures partial decoupling
of resources; if and when a resource is created for a given stack,
any dependent resources (with the exception of s3 buckets) will also
be created within and further associated with that stack.