Introduce script for deploying pipelines
This commit introduces a script, deploy-aws-pipes, to help in deploying
CodePipeline pipelines. The script uses the aws cli to deploy the
exodus-pipeline.yaml template for dev, stage, and prod environments.

The configuration file used by the script is necessary for storing
private values. As such, it is not provided by this repository.
Additionally, any other variables set in the configuration file will
override the script's defaults and any options/parameters provided.