Use real branch name in verify-source [RHELDST-1288]
This commit changes verify-source's REPO_BRANCH variable to the branch
name from the source stage rather than the template parameter value.

This resolves an issue wherein no branch is specified by the user,
preferring to instead use the template's default values, and an attempt
to checkout "None" as a branch is made.