Parameterize whitelist used for verify-source [RHELDST-1290]
Previously initial stage pipeline runs were guaranteed to fail on the
'verify' action due to missing whitelist in the pipeline's artifact
bucket. This commit implements a new resource to provision the bucket
with an initial whitelist.

If a URL and lambda function role ARN are passed to the
exodus-pipeline-resources template, a lambda function is created and
run, uploading the content at the URL to the pipeline artifact bucket.

Note: URLs should point to publicly accessible raw/plain text, e.g.,
https://raw.githubusercontent.com/.../.whitelist.gpg.