exodus-lambda requires cdn-definitions-private [RHELDST-4703]
The configuration that was once available in the cdn-definitions
PyPI package has since moved to a private git repository:
"cdn-definitions-private".

In order for exodus-lambda to consume data from a private repo,
a "GITHUB_TOKEN" environment variable must be set. GITHUB_TOKEN
should be assigned a GitHub access token with the "repo" scope,
which enables clients (e.g., pip) access to private repos.