Fix invalid MIME type for RPMs
Fixes an annoyance which has been present for a long time. The
sample RPMs in the docs and the reference test data used
"audio/x-pn-realaudio-plugin" as the content type.

While technically accurate in that the current CDN does actually serve
them with this type, it's clearly for no good reason - the legacy
backend presumably has some naive mapping between file extensions and
MIME types, and for ".rpm", this wrong type happens to have higher
priority.

It's not a behavior we should carry over to exodus, and having this
example in the docs is confusing at best, so let's fix it up.