Consume cdn-definitions from table [RHELDST-7898]
The exodus-lambda functions previously consumed the cdn definitions
from the the exodus-lambda package. This workflow required the lambdas
to be rebuilt and redeployed after each update to the cdn-definitions
data set.

Release engineers should be able to update the cdn-definitions data
set without rebuilding and redeploying the lambdas. The definitions
are now stored in a $project-config-$env DynamoDB table, which is
decoupled from the exodus-lambda package.