reftest: add progress bar for uploads
This script performs both downloads and uploads for some large-ish
files. We previously had a progress bar for downloads, but not
uploads.

For me at least the uploads can be slow enough to give the impression
that the script is stuck, so let's add a progress bar for uploads too.