Fix condition in lambda deploy template
Previously, the EnableKeyGroup condition in exodus-lambda-deploy.yaml
checked that a joined list of keyids was an empty string. However, the
default value of keyids is the string "None", so the joined keyids
string will never be "".

The failure of this condition meant that an attempt was made to create
a KeyGroup resource even when no valid IDs existed.

This commit corrects the condition to look for the string "None".