Split config file generation into a separate script
We do env var substitution on a template in order to generate
lambda_config.json. This was previously embedded in
scripts/build-package.

It's useful sometimes to be able to do that outside of building the
package, e.g. in order to manually generate a config file for
testing, so let's move that into a separate script.