Allow AWS endpoint URL to be customized
If EXODUS_AWS_ENDPOINT_URL env var is set, we'll use that as the AWS
endpoint URL when constructing boto clients.

This will never be used in production. The motivation is only to enable
testing against an instance of localstack.

The same applies to the reftest support script, as that also works fine
against localstack if the endpoint URL is set. Though reftest has
command-line arguments, the same env var is used there for consistency.