Introduce fakefront: supporting dev server
Run `tox -e fakefront' to deploy lambdas via a local development server.
This server emulates enough of the CloudFront Lambda@Edge behaviors for
the exodus-lambda code to be usable. It is primarily intended to be used
with the localstack container deployed via the exodus-gw dev env.

Possible uses of this include:

- more productive development: test local changes instantly with hot
  reload, don't need to do a new lambda deployment for every change

- work with a fully functional exodus-rsync => exodus-gw =>
  exodus-lambda chain locally without involving any remote services

- deploy a fully functional integration test environment in openshift
  without having to deploy any AWS resources / incur any AWS costs