Make more config items fully parametrizable
Allow a few more elements in lambda_config.template to be parametrized
by environment variables.

In the case of the DynamoDB table names, it's
better if the full name can be parametrized because not everything
necessarily matches the $PROJECT-cdn-$ENV_TYPE pattern (example: the
tables deployed by exodus-gw dev env don't).

In the case of the log format, it's useful to be able to switch the
format so that %(aws_request_id)s is avoided, as that variable won't be
available if the code runs outside the context of Lambda@Edge.

The default values of everything remains the same, except that an
unnecessary "\n" was removed from the end of the log format.