Add "X-Exodus-Query" header [RHELDST-9321]
Our tests sometimes seem to hit an old version of our lambda, for
unclear reasons. Let's make it possible to query the currently deployed
lambda version so we can at least reliably detect when this happens.

The "X-Exodus-Query" header is primarily used in a request, to provide
the version of exodus-lambda, "X-Exodus-Version" in response.