refactor: support reading config from env vars directly [RHELDST-11210]
We currently read config from a lambda_config.json which is included in
the lambda deployment package along with the code.

A problem with this is that we can't have any differences in config
between environments while simultaneously ensuring that we deploy
exactly the code in a later environment which we've validated in an
earlier environment, since code and config are bundled together.

We can fix this by ditching lambda_config.json and just reading
environment variables directly, which will have to be configured on the
created lambda resources.

This commit implements part 1 of the migration to env vars: we support
loading config from env vars, while still also supporting
lambda_config.json. These will both be supported for a brief
transitional period while ansible playbooks are updated to move towards
the env var approach.

The lambda cloudformation template adds a new parameter,
lambdaenvironmenturi. This is not yet used but was added so that ansible
can start passing it. Later, env vars will be loaded from that URI via
an AWS::Include transform when the template is evaluated.