Ensure URL always ends with / when serving index [RHELDST-13170]
When trying to emulate directories by serving up index pages for certain
URIs it is important that the browser has requested the URL of the
"directory" while including a trailing /. If not, any relative links
in an HTML index will not resolve as desired.
We therefore need to send a redirect to the requested URL with /
appended if it wasn't already present.

As part of this fix, the autoindex test also had to be fixed as it was
not actually covering autoindex. It had mocked the DynamoDB query to
always return a result, but in reality the autoindex scenario first does
one query which fails to find a result, and then a second query to look
up the autoindex.