Add a support script to generate signed URLs/cookies from command-line
Something like this has been floating around JIRA comments for a while
and used during testing. Add a proper script to handle it because we'll
have an ongoing need to generate these signatures.

The script includes a self-test feature to ensure that trying to use the
wrong key or ID for an environment will be detected.