Add missing Set-Cookie attributes for ESS compliance [RHELDST-14938]
ESS guideline SEC-APP-REQ-1 requires cookies be set with a number of
attributes. This commit adds the missing attributes, HttpOnly and Domain.
SameSite is omitted, left to the default of 'lax', as the cookie is set
by lambda but used by cloudfront. Domain is set to the distribution named
in the event request's config.