reftest: support publishing via exodus-gw
The reftest script previously worked only by accessing S3 and DynamoDB
directly. We really ought to avoid that as much as possible, especially
on stage and production environments. Ideally we should not even have
the necessary credentials for that.

Let's support deploying reftest data via exodus-gw. That's safer, will
be logged appropriately and so on.

The primary motivation here is to make it possible to more safely
provision reftest data to newly deployed stage and production
environments, though a few other changes relating to safety were made
as well:

- don't deploy "absent" items - there is no need to, and it kinda
  defeats the purpose. e.g. the reftest entry for "does-not-exist.iso"
  is supposed to test what happens if you query content which doesn't
  exist... not if you query content which has been explicitly marked
  absent, which is subtly different.

- don't deploy exodus-config by default. I don't think this should have
  been included in the default reftest steps. Most of the reftest data
  represents some invariants which should always be true and should be
  safe to deploy on a live environment. exodus-config is different in
  that the config is expected to change over time, and deploying a wrong
  config to a live environment can break the CDN completely.

- don't accept a --release-date argument. exodus-gw doesn't support it,
  and I see no reason to let the user control this anyway; there is no
  use-case for deploying reftest data in the future, and allowing the
  user to provide this argument opens the possibility of them providing an
  invalid argument (e.g. timestamps of the incorrect format, or strings
  which are not valid timestamps at all).