reftest: implement caching
Not sure how it is for others, but for me downloading some of the larger
files in reftest data is significantly slow. For files with a known
checksum, if we've already downloaded the data once it could obviously
be cached and reused at a later run, so let's do that.

This is especially useful if reftest data needs to be deployed to
multiple environments in a row, as otherwise the script would waste time
redownloading files for every environment.