logging: ensure JSON logs include logger name
When we moved to JSON logging, the format adopted did not include the
logger's name anywhere. That's a problem since, if we can't tell which
logger produced which message, it'll sometimes be impossible to tell
where a message came from and we will not be able to decide which
loggers to enable/disable.

Change it to ensure logger name appears in a field, as in example:

    {"level": "DEBUG",
     "logger": "botocore.retryhandler",
     "time": "2023-05-23 23:30:19.194",
     "aws-request-id": null,
     "message": "No retry needed.",
     "request": null,
     "response": null}