docs: define and document a config file schema
We should define a schema for the config file. The primary motivation is
to make it possible to validate the config file during lambda
build/deployment, so that we won't waste time deploying a lambda package
which immediately crashes due to bad config. Recent type errors as
mentioned e.g. in 55c2d2275085 demonstrate the need for validation.