Refactor/fix multi-region handling [RHELDST-15297]
This commit reworks the handling of multiple regions for DynamoDB
tables. We have long had the ability to configure multiple available
regions, but the meaning of this region list has been changed.

Previously: this was a list of all regions expected to contain replicas,
and the lambda would use this to decide if it could consult a table in
its own local region for reduced latency.

Now: this is a list of all regions expected to contain replicas, *in the
order in which they should be consulted regardless of the execution
region*. Queries are tried on each region in order until one of them
succeeds. In other words, we have a single-primary setup [1] where the
first listed region is the primary, and all others are secondaries which
are used only for failover to improve resilience.

The main reason for this change is the realization that we need strongly
consistent reads. The prior setup has never been used in practice because
we didn't deploy multi-region tables yet, but we are about to do so.

Also, the handling of multiple regions was buggy in some ways which have
also been fixed. Those bugs were hidden until now since we only used a
single region in practice.

- there were 'available_regions' in config for both the 'table' and
  'config_table', yet the list for 'config_table' was simply ignored as
  no code ever accessed it.

- the 'available_regions' for table was also used to decide which region
  should be used for Secrets Manager. It doesn't make sense for those to
  be tied together since replication of DynamoDB and Secrets Manager are
  unrelated, and there are no plans to have multi-region SM. If
  available_regions had ever been set to anything other than
  ['us-east-1'], this would have caused crashes. The SM region can have
  its region configured independently.

[1] https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/bp-global-table-design.prescriptive-guidance.writemodes.html#bp-global-table-design.prescriptive-guidance.writemodes.single-primary