Fix trailing / redirect leaking alias info [RHELDST-21903]
In order for directory listings to work properly for users, if the
user requests "/some/path", they will be redirected to the same path
with a trailing slash added, i.e. "/some/path/".

Problem: the implemented logic was previously returning a path to the
user *after* the resolution of rhui and releasever aliases.
For example:

- user requested: /content/dist/rhel9/rhui/9/x86_64/supplementary/os
- redirected to:  /content/dist/rhel9/9.3/x86_64/supplementary/os/

The usage of aliases should not be leaked in this way. Fix it to work as
desired, which is for example:

- user requested: /content/dist/rhel9/rhui/9/x86_64/supplementary/os
- redirected to:  /content/dist/rhel9/rhui/9/x86_64/supplementary/os/