Fix latest requirements not usable for deployment
We ran pip-compile here without any particular arguments, while in the
final deployment environment we run it with arguments selecting binary
packages (wheels) only.

frozendict 2.4.0 exposed that this is flawed: there is no cp311 or py3
wheel uploaded for that at time of writing. This means that release is
unusable during deployment for us as we require a compatible wheel; but
since pip-compile was allowed to use source packages, it would consider
2.4.0 as a usable version.

To avoid such issues we must ensure that both pip-compile here and 'pip
install' during deployment use compatible package-selection args. Update
the pip-compile commands here accordingly.

Additionally, with the new args, package resolution for
test-requirements failed on 'coveralls'. We don't use that anymore, so
just remove it.