Strip certain headers from S3 responses [RHELDST-22613]
We were allowing all headers returned by S3 to propagate normally.
It would be better to strip some of them.
These are considered to be internal implementation details and we'd
rather not have clients come to depend on them, since that might make
it difficult for us to change things later. For example: we don't
want people to start extracting MD5 checksums from
"x-amz-meta-exodus-migration-md5".

The set of headers to strip can be specified by a new 'strip_headers'
config item, but there's a reasonable hardcoded default.