Support compressed configuration [RHELDST-25461]
Config has formerly always been stored as a String attribute on DynamoDB
items. Let's support also loading it as a gzip-compressed Binary
attribute.

The motivation is to allow more room for the config to grow.
The max DynamoDB item size is 400KB, and a realistic string-based config
has already reached 70KB (about 18% of the limit). As the config is only
ever expected to grow over time, we may eventually hit that limit without
warning; this discourages us from adding more fields into the config.

The same config compressed uses about ~5KB (about 1% of the limit), which
seems small enough that we'll never have to worry about hitting the
limit and we can freely add new fields into the config.

This change should be fully deployed to all exodus-cdn environments
before deploying the corresponding change to exodus-gw.