Fix double base64 encoding in config [RHELDST-25461]
It was not correct to do a base64 decode here, because botocore already
does that implicitly on both reads and writes of 'B' attributes on
DynamoDB items. Drop the unnecessary decoding.

Note that the code here "worked" because the writing side in exodus-gw
made the same mistake, ultimately ending up with items having two layers
of base64 encoding. That will be fixed, but the reading side here needs
to be fixed first.