Introduce mirrored reads [RHELDST-28332]
This commit modifies exodus-cdn's lookup behavior to enable mirrored
reads. This means, that when exodus-cdn looks up content to be served
for a path with a $releasever alias in effect, exodus-cdn attempts to
look up content on both sides of the alias. (Previously, exodus-cdn
only performed a content lookup on the destination side.)

The mirrored reads are enabled by default, but they can be disabled
by setting the new config option "mirrored_reads" to false (0, '0',
or 'False', where 'False' is not case sensitive.)